+++ modules/core/tests/server/core.server.config.tests.js
@@ -30,9 +30,9 @@ describe('Configuration Tests:', function () {
           username: 'user_config_test',
           provider: 'local',
           email: 'user_config_test_@localhost.com',
-          firstName: 'User',
-          lastName: 'Local',
-          displayName: 'User Local',
+          firstname: 'User',
+          lastname: 'Local',
+          displayname: 'User Local',
           roles: ['user']
         };
 
@@ -40,10 +40,10 @@ describe('Configuration Tests:', function () {
           username: 'admin_config_test',
           provider: 'local',
           email: 'admin_config_test_@localhost.com',
-          firstName: 'Admin',
-          lastName: 'Local',
-          displayName: 'Admin Local',
-          roles: ['user', 'admin']
+          firstname: 'Admin',
+          lastname: 'Local',
+          displayname: 'Admin Local',
+          roles: ['user', 'admin', 'seller']
         };
 
         userFromSeedConfig = config.seedDB.options.seedUser;