+++ modules/users/server/config/strategies/paypal.js
@@ -25,9 +25,9 @@ module.exports = function (config) {
 
     // Create the user OAuth profile
     var providerUserProfile = {
-      firstName: profile.name.givenName,
-      lastName: profile.name.familyName,
-      displayName: profile.displayName,
+      firstname: profile.name.givenName,
+      lastname: profile.name.familyName,
+      displayname: profile.displayname,
       email: profile._json.email,
       username: profile.username,
       provider: 'paypal',