+++ modules/users/server/controllers/admin.server.controller.js
@@ -22,9 +22,9 @@ exports.update = function (req, res) {
   var user = req.model;
 
   //For security purposes only merge these parameters
-  user.firstName = req.body.firstName;
-  user.lastName = req.body.lastName;
-  user.displayName = user.firstName + ' ' + user.lastName;
+  user.firstname = req.body.firstname;
+  user.lastname = req.body.lastname;
+  user.displayname = user.firstname + ' ' + user.lastname;
   user.roles = req.body.roles;
 
   user.save(function (err) {
@@ -59,7 +59,7 @@ exports.delete = function (req, res) {
  * List of Users
  */
 exports.list = function (req, res) {
-  User.find({}, '-salt -password').sort('-created').populate('user', 'displayName').exec(function (err, users) {
+  User.find({}, '-salt -password').sort('-created').populate('user', 'displayname').exec(function (err, users) {
     if (err) {
       return res.status(400).send({
         message: errorHandler.getErrorMessage(err)